+++ crates/raven-database/src/database/sqlite/query/mod.rs
@@ -1,3 +1,4 @@
+use delete::DeleteStatement;
 use insert::InsertStatement;
 use select::SelectStatement;
 use update::UpdateStatement;
@@ -5,6 +6,7 @@ use update::UpdateStatement;
 mod insert;
 mod select;
 mod update;
+mod delete;
 
 #[derive(Debug, Clone)]
 /// Shorthand for constructing any table query
@@ -24,6 +26,11 @@ impl Query {
     /// Construct a table [`InsertStatement`]
     pub fn insert<'a>() -> InsertStatement<'a> {
         InsertStatement::new()
+    }
+
+    /// Construct a table [`DeleteStatement`]
+    pub fn delete<'a>() -> DeleteStatement<'a> {
+        DeleteStatement::new()
     }
 }